requirements.txt
uvicorn==0.23.2
pydantic==2.5.0
faiss-cpu==1.7.4
langchain-together==0.0.2.post1
langserve[server]>=0.0.30
langchain-cli>=0.0.15
pinecone-client==2.2.4
unstructured==0.12.0
